SAAS-31 give some SRP sanity to pmm-inventory
