+++ pmm-app/src/pmm-inventory/Tabs/index.ts
@@ -1,0 +1,3 @@
+export { AgentsTab } from './AgentsTab';
+export { NodesTab } from './NodesTabs';
+export { ServicesTab } from './ServicesTab';
