+++ pmm-app/src/pmm-inventory/module.tsx
@@ -1,7 +1,7 @@
 import { PanelPlugin } from '@grafana/data';
+import { InventoryPanel } from './panel';
 import '../react-plugins-deps/styles.scss';
 import '../react-plugins-deps/style.less';
 import './panel.scss';
-import InventoryPanel from './panel';
 
 export const plugin = new PanelPlugin(InventoryPanel);
