+++ pmm-app/src/pmm-inventory/panel.test.tsx
@@ -3,7 +3,7 @@ import { Table } from 'antd';
 import { render, unmountComponentAtNode } from 'react-dom';
 import { act } from 'react-dom/test-utils';
 import { InventoryDataService } from './DataService';
-import { agentsColumns, nodesColumns, servicesColumns } from './panel';
+import { agentsColumns, nodesColumns, servicesColumns } from './panel.constants';
 
 jest.mock('../react-plugins-deps/components/helpers/notification-manager', () => () => ({
   showErrorNotification: () => {},
