+++ pmm-app/src/react-plugins-deps/style.less
@@ -1,5 +1,3 @@
-//@import '../../node_modules/antd/dist/antd.css';
-
 #antd {
   background-color: transparent;
   @import '../../node_modules/antd/dist/antd.less';
@@ -17,8 +15,8 @@
 @font-size-base: 14px; // major text font size
 @heading-color: white; // heading text color
 @text-color: white; // major text color
-@text-color-secondary : white; // secondary text color
-@disabled-color : rgba(0, 0, 0, .25); // disable state color
+@text-color-secondary: white; // secondary text color
+@disabled-color: rgba(0, 0, 0, 0.25); // disable state color
 @border-radius-base: 0px; // major border radius
 @border-color-base: rgb(40, 40, 40); // major border color
 @box-shadow-base: 0 2px 8px rgba(0, 0, 0, 0.15); // major shadow for layers
